removed DelegatedGrantMessage circular dependency
